Fix TODO items 453-470: shader errors, bone init, time uniform, GPU resource leaks, draw call type, animation flags, bone remap table
- TODO 453: Verified — unpopulated struct fields never read by any code
- TODO 454: Verified — structurally different but functionally equivalent
- TODO 455: Fixed — compileShaders() now throws on error like JS
- TODO 456: Verified — documented intentional difference
- TODO 457: Verified — matches JS behavior exactly
- TODO 458: Verified — ImGui immediate-mode doesn't need Vue watchers
- TODO 459: Verified — setSlotFileLegacy is correct for string paths
- TODO 460: Fixed — removed extra getTextureFile() fallback not in JS
- TODO 461: Fixed — boneless bone_matrices now all zeros like JS
- TODO 462: Fixed — u_time uses app-relative time like performance.now()
- TODO 463: Fixed — added glDeleteBuffers() before clearing
- TODO 464: Verified — geoset_mapper::map only modifies labels
- TODO 465: Fixed — M2LegacyDrawCall::count now uint32_t
- TODO 466: Verified — ImGui doesn't need Vue watchers
- TODO 467: Fixed — stopAnimation() clears anim source flags
- TODO 468: Fixed — boneless bone_matrices now all zeros like JS
- TODO 469: Verified — direct CASC access functionally equivalent
- TODO 470: Fixed — buildBoneRemapTable uses m2 or skel bones

Agent-Logs-Url: https://github.com/NathanNatNat/wow.export.cpp/sessions/a8cb88f1-986e-490e-b588-1930df07c0da

+++ src/js/3D/renderers/M2LegacyRendererGL.h
@@ -42,7 +42,7 @@ static constexpr int LEGACY_PIXEL_SHADER = 0;  // Combiners_Opaque
 struct M2LegacyDrawCall {
 	gl::VertexArray* vao = nullptr;
 	uint32_t start = 0;
-	uint16_t count = 0;
+	uint32_t count = 0;
 	std::array<int, 4> tex_indices = { -1, -1, -1, -1 };
 	uint16_t texture_count = 1;
 	int vertex_shader = LEGACY_VERTEX_SHADER;
